chore: Table Promotion from preview
