+++ modules/preview-react/_examples/stories/TablesAdvanced.stories.mdx
@@ -1,4 +1,4 @@
-import {Table} from '@workday/canvas-kit-preview-react/table';
+import {Table} from '@workday/canvas-kit-react/table';
 
 import {SelectableRows} from './examples/Table/WithSelectableRows';
 import {ExpandableRows} from './examples/Table/WithExpandableRows';
