--- modules/preview-react/table/README.md
@@ -1,5 +1,0 @@
-# Canvas Kit Table
-
-View the
-[documentation for Table](https://workday.github.io/canvas-kit/?path=/docs/preview-table--basic) on
-Storybook.
