--- modules/preview-react/table/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/Table';
