--- modules/preview-react/table/spec/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../../../tsconfig.spec.json"
-}
