--- modules/preview-react/table/stories/examples/Basic.tsx
@@ -1,36 +1,0 @@
-import React from 'react';
-
-import {Table} from '@workday/canvas-kit-preview-react/table';
-
-export const Basic = () => {
-  return (
-    <Table>
-      <Table.Caption>Coffee Drinks and Sizes</Table.Caption>
-      <Table.Head>
-        <Table.Row>
-          <Table.Header scope="col" backgroundColor="soap100">Drink</Table.Header>
-          <Table.Header scope="col" backgroundColor="soap100">Size</Table.Header>
-        </Table.Row>
-      </Table.Head>
-      <Table.Body>
-        <Table.Row>
-          <Table.Cell>Espresso</Table.Cell>
-          <Table.Cell>1 oz</Table.Cell>
-        </Table.Row>
-        <Table.Row>
-          <Table.Cell>Macchiato</Table.Cell>
-          <Table.Cell>2 oz Espresso</Table.Cell>
-        </Table.Row>
-        <Table.Row>
-          <Table.Cell>Cortado</Table.Cell>
-          <Table.Cell>2 oz Espresso, 1 oz Foamed Milk</Table.Cell>
-        </Table.Row>
-        <Table.Row></Table.Row>
-        <Table.Row>
-          <Table.Cell>Cappuccino</Table.Cell>
-          <Table.Cell>2 oz Espresso, 2 oz Foamed Milk, 2 oz Steamed Milk</Table.Cell>
-        </Table.Row>
-      </Table.Body>
-    </Table>
-  );
-};
