--- modules/preview-react/table/stories/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../../../tsconfig.stories.json"
-}
