+++ modules/react/table/README.md
@@ -1,7 +1,5 @@
 # Canvas Kit Table
 
 View the
-[documentation for Table](https://workday.github.io/canvas-kit/?path=/docs/components-containers-table-react--basic)
+[documentation for Table](https://workday.github.io/canvas-kit/?path=/story/components-containers-table--basic)
 on Storybook.
-
-[> Workday Design Reference](https://design.workday.com/components/containers/tables)
