+++ modules/react/table/lib/TableFooter.tsx
