+++ modules/react/table/stories/Table.stories.mdx
@@ -1,25 +1,19 @@
 import {SymbolDoc} from '@workday/canvas-kit-docs';
-
-import {Table, TableRow} from '@workday/canvas-kit-react/table';
-import {StatusIndicator} from '@workday/canvas-kit-preview-react/status-indicator';
-
+import {Table} from '@workday/canvas-kit-react/table';
+// Examples
 import {Basic} from './examples/Basic';
-import {RowHeader} from './examples/RowHeader';
-import {RowStates} from './examples/RowStates';
+import {BasicWithHeading} from './examples/BasicWithHeading';
+import {FixedColumn} from './examples/FixedColumn';
+import {RightToLeft} from './examples/RightToLeft';
 
 <Meta title="Components/Containers/Table" component={Table} />
 
 # Canvas Kit Table
 
-<StatusIndicator variant="red">
-  <StatusIndicator.Label>Deprecated</StatusIndicator.Label>
-</StatusIndicator>
-
-`Table` has been deprecated and will be removed in a future major version. Please use [`Table`](https://workday.github.io/canvas-kit/?path=/docs/preview-table--basic) in Preview instead.
-
-Tables are an efficient way of displaying sets of repeating data with the same structure.
-
-[> Workday Design Reference](https://design.workday.com/components/containers/tables)
+`Table` is a simple styled compound component that renders a
+[table](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/table) element. It is used to
+present information in a two-dimensional table comprised of rows and columns of cells containing
+data.
 
 ## Installation
 
@@ -31,76 +25,43 @@ yarn add @workday/canvas-kit-react
 
 ### Basic Example
 
-Table includes a container `Table` component and a `TableRow` component. `Table` is a simple wrapper
-of the `<table>` element with Canvas styles. `TableRow` wraps `<tr>` and supports additional props
-to customize the style of the row.
+Users may not want to use a `caption` so they can import [Heading](/components/text/heading/) or
+[Text](/components/text/text/) instead. This will give the user more flexibility around the
+customization of the title/heading of their table.
 
-Use `Table` and `TableRow` in conjunction with the standard `<thead>` and `<tbody>` elements to
-create your Table.
+<ExampleCodeBlock code={BasicWithHeading} />
+
+### Right to Left
+
+Table supports right-to-left languages when specified in the CanvasProvider theme.
+
+<ExampleCodeBlock code={RightToLeft} />
+
+### Example with Caption
+
+Users are free to use a `caption` instead of a heading. A `caption` is not required but it is good
+for
+[accessibility](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/table#accessibility_concerns)
+purposes.
 
 <ExampleCodeBlock code={Basic} />
 
+### Fixed Column
+
+Users may add styles to the `Table.Header` to render a fixed column. The example below assigns a
+`width` to the `Table` to guarantee the fixed column is triggered, but you are free to omit the
+`width` if you would only like the fixed column to be triggered if necessary.
+
+<ExampleCodeBlock code={FixedColumn} />
+
+### Advanced
+
+You can also find several advanced Table examples in our Storybook Examples section.
+
+- [Expandable Rows](/docs/examples-tables-advanced--expandable-rows)
+- [Selectable Rows ](/docs/examples-tables-advanced--selectable-rows)
+- [Sortable Column Headers](/docs/examples-tables-advanced--sortable-column-headers)
+
 ## Component API
 
-### Table
-
-`Table` is a simple wrapper of the `<table>` element with Canvas styles. Use it as you would a
-standard `<table>` element.
-
-```tsx
-<Table>
-  <thead>
-    <TableRow header={true}>
-      <th>ID</th>
-      <th>Name</th>
-      ...
-    </TableRow>
-  </thead>
-  <tbody>
-    <TableRow>
-      <td>1</td>
-      <td>Aidan Brown</td>
-      ...
-    </TableRow>
-    ...
-  </tbody>
-</Table>
-```
-
-#### Props
-
-`Table` does not have any documented props. Undocumented props are spread to its underlying
-`<table>` element.
-
-### TableRow
-
-`TableRow` wraps `<tr>` and supports additional props to customize the style of the row. Use it as
-you would a standard `<tr>` element.
-
-Set the `header` prop of the `TableRow` to designate whether or not it contains header elements.
-
-<ExampleCodeBlock code={RowHeader} />
-
-Set the `state` prop of the `TableRow` to display it using one of several different visual states.
-`state` accepts the following values:
-
-- `TableRow.State.Selected`
-- `TableRow.State.Alert`
-- `TableRow.State.InputAlert`
-- `TableRow.State.Error`
-- `TableRow.State.InputError`
-
-| State                       | Purpose                                                                |
-| --------------------------- | ---------------------------------------------------------------------- |
-| `TableRow.State.Alert`      | When there is an issue with the entire row as a whole.                 |
-| `TableRow.State.InputAlert` | When there are issues with one or more (but not all) cells in the row. |
-| `TableRow.State.Error`      | When there is an error with the entire row as a whole.                 |
-| `TableRow.State.InputError` | When there are errors with one or more (but not all) cells in the row. |
-| `TableRow.State.Selected`   | Indicates the selected row to the user.                                |
-
-<ExampleCodeBlock code={RowStates} />
-
-`InputAlert` and `InputError` are identical to the `Alert` and `Error` states, respectively, except
-they lack a darkened border around the row.
-
-<SymbolDoc name="TableRow" fileName="/react/" />
+<SymbolDoc name="Table" fileName="/react/" />
