+++ modules/react/table/stories/examples/BasicWithHeading.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import {Table} from '@workday/canvas-kit-preview-react/table';
+import {Table} from '@workday/canvas-kit-react/table';
 import {Heading} from '@workday/canvas-kit-react/text';
 import {useUniqueId} from '@workday/canvas-kit-react/common';
 
 export const BasicWithHeading = () => {
   const headingID = useUniqueId();
-  
+
   return (
     <>
       <Heading as="h3" id={headingID} size="small" marginBottom="s">
@@ -15,8 +15,12 @@ export const BasicWithHeading = () => {
       <Table aria-labelledby={headingID}>
         <Table.Head>
           <Table.Row>
-            <Table.Header scope="col" backgroundColor="soap100">Toppings</Table.Header>
-            <Table.Header scope="col" backgroundColor="soap100">Amount</Table.Header>
+            <Table.Header scope="col" backgroundColor="soap100">
+              Toppings
+            </Table.Header>
+            <Table.Header scope="col" backgroundColor="soap100">
+              Amount
+            </Table.Header>
           </Table.Row>
         </Table.Head>
         <Table.Body>
